Put stack length variable back to ulong

+++ src/nel/parse/parser.d
@@ -163,7 +163,7 @@ class Parser
                 else
                 {
                     // Pop previous scanner off stack.
-                    uint top = includeStack.length - 1;
+                    ulong top = includeStack.length - 1;
                     scanner = includeStack[top];
                     includeStack.length = top;
                     // Ready a new token for the scanner.
